feat(ui): updated create secret form in overview to better ux
