+++ frontend/src/views/SecretOverviewPage/components/CreateSecretForm/index.tsx
@@ -1,0 +1,1 @@
+export { CreateSecretForm } from "./CreateSecretForm"
